fix: ensure HTTPS instance is destroyed at server shutdown
