+++ config/config.test.json.sample
@@ -4,7 +4,7 @@
   },
   "server": {
     "host": "127.0.0.1",
-    "port": 8000
+    "port": 5000
   },
   "api": {
     "host": "127.0.0.1",
